Refactor javascript to use ES6 modules and remove event handlers from HTML.
This involved creating a components directory which holds an exported
class in each file that represents functionality of a certain set of
elements on the page. There is now a main.js that acts as a router
depending on the current page. The current page is marked by data
attribute in the first div of the <main> element.